Separate __main__.py and ServerSideAPI
